修改版本和代码规范

+++ rayin-design-server/src/main/java/ink/rayin/app/web/cache/RedisTemplateUtil.java
@@ -19,7 +19,6 @@ import static org.springframework.data.redis.connection.ReturnType.INTEGER;
  */
 @Component("redisTemplateUtil")
 public class RedisTemplateUtil{
-    //注入StringRedisTemplate
     @Resource
     StringRedisTemplate stringRedisTemplate;
 
@@ -197,7 +196,7 @@ public class RedisTemplateUtil{
             }
         });
         RedisConnectionUtils.unbindConnection(stringRedisTemplate.getConnectionFactory());
-        return res;
+        return res.longValue();
     }
 
     /**
